Implemented animations and FMOD OnBeatTimelineEvent
Slam and JumpSlam animations are in! They still need to be fixed in terms of dealing damage.

Also added FMODBanks folder to the repo so that everyone can work on the same file in FMOD!

JumpSlam is also kinda buggy in that the Tuba overshoots it's target and lands farther than intended.

OnBeatTimelineEvent now works too, TAKE NOTE: you need to set a tempo marker for it to be done on beat.