Made playing tremolo more consistent.
Adjusted firing tremelo so that it is more consistent.
Also made it so the tremelo is the only projectile that can destroy the tuba's ranged projectiles.

Also here are the new projectile controls: Rapidly press the right trigger only to fire the tremelo, and press and hold the left trigger to fire the beam.