Added a dedicated playstyle switch button, mapped to F on Keyboard (top face button on controller)
Implemented a dedicated button to switch playstyles. It's a lil busted due to no buffer time of switching playstyles (causes the known bug to play both styles at the same time, and DESYNCS the music...) I also implemented an equip system for playstyle loadouts later down the line, and it's proving to help a ton with modularity. Also touched up on the cloth physics of Maestro's tailcoat.