More bug fixes
Adjusted win screen to transition nicely if you time it correctly.
Changed the ZOrder of the pause screen to hide the HUD.
Enabled gamepad support for the win and lose screens.
- Changed them from User Widgets to Common Activatable Widgets
Adjusted Slam Hitbox and tracking for Tuba.
- Should no longer be able to safely hide between Tuba's legs.
Adjusted lock on for weird slowdown when moving aim around.
Fixed blend space for slowly walking left.