Messing around in FMOD for Pizza Bomb
(Did somebody order a pizza bomb?)
Used transition region w/ condition parameters to have the sfx play on beat.
Also added a nested event that incluces the pizzicato multi-instrument so that the game can detect when the music starts playing to spawn the bomb via the IsEventValid blueprint node in UE5.